requirements.txt
# 量化回测系统依赖包
# 仅包含当前代码中实际使用的依赖

# 核心数据处理
pandas>=1.5.0

# 配置管理
PyYAML>=6.0

# 数据源
pywencai>=0.2.0